rename snapshot_specific.hpp -> snapshot_detail.hpp.
