+++ libraries/chain/block_header_state_legacy.cpp
@@ -1,6 +1,6 @@
 #include <eosio/chain/block_header_state_legacy.hpp>
 #include <eosio/chain/block_header_state_utils.hpp>
-#include <eosio/chain/snapshot_specific.hpp>
+#include <eosio/chain/snapshot_detail.hpp>
 #include <eosio/chain/exceptions.hpp>
 #include <limits>
