+++ libraries/chain/block_state_legacy.cpp
@@ -1,7 +1,7 @@
 #include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain/block_header_state_utils.hpp>
 #include <eosio/chain/exceptions.hpp>
-#include <eosio/chain/snapshot_specific.hpp>
+#include <eosio/chain/snapshot_detail.hpp>
 
 
 namespace eosio::chain {
