+++ libraries/chain/include/eosio/chain/snapshot_detail.hpp
